feat(ui): Refactoring: Create Identifier (#828)
* feat(ui): implement create identifier

* feat(ui): implement group identifier

* feat(ui): update all component use create identifier

---------
