+++ src/ui/components/CreateGroupIdentifier/components/GroupMembers.test.tsx
@@ -1,5 +1,5 @@
 import { setupIonicReact } from "@ionic/react";
-import { mockIonicReact } from "@ionic/react-test-utils";
+import { ionFireEvent, mockIonicReact } from "@ionic/react-test-utils";
 import { fireEvent, render, waitFor } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import { Provider } from "react-redux";
@@ -7,23 +7,24 @@ import configureStore from "redux-mock-store";
 import EN_TRANSLATIONS from "../../../../locales/en/en.json";
 import { store } from "../../../../store";
 import { connectionsFix } from "../../../__fixtures__/connectionsFix";
-import { IdentifierStage3 } from "./IdentifierStage3";
-import { IdentifierColor } from "./IdentifierColorSelector";
+import { GroupMembers } from "./GroupMembers";
+import { IdentifierColor } from "../../CreateIdentifier/components/IdentifierColorSelector";
+import { Stage } from "../CreateGroupIdentifier.types";
 
 setupIonicReact();
 mockIonicReact();
 
-describe("Identifier Stage 3", () => {
+describe("Identifier Stage 2", () => {
   const mockStore = configureStore();
 
-  const stage3State = {
-    identifierCreationStage: 3,
+  const stage2State = {
+    identifierCreationStage: 0,
     displayNameValue: "",
     selectedAidType: 0,
     selectedTheme: 0,
     threshold: 1,
-    scannedConections: [],
-    selectedConnections: [connectionsFix[0]],
+    scannedConections: [connectionsFix[0]],
+    selectedConnections: [],
     ourIdentifier: "",
     newIdentifier: {
       id: "",
@@ -54,10 +55,10 @@ describe("Identifier Stage 3", () => {
   const resetModal = jest.fn();
 
   test("Renders default content", async () => {
-    const { getByTestId, getByText } = render(
+    const { getByText } = render(
       <Provider store={storeMocked}>
-        <IdentifierStage3
-          state={stage3State}
+        <GroupMembers
+          state={stage2State}
           setState={setState}
           componentId={"create-identifier-modal"}
           setBlur={setBlur}
@@ -66,26 +67,62 @@ describe("Identifier Stage 3", () => {
       </Provider>
     );
     expect(
-      getByText(EN_TRANSLATIONS.createidentifier.threshold.title)
+      getByText(EN_TRANSLATIONS.createidentifier.connections.title)
+    ).toBeVisible();
+    expect(getByText(connectionsFix[0].label)).toBeVisible();
+    expect(
+      getByText(EN_TRANSLATIONS.createidentifier.connections.continue)
     ).toBeVisible();
     expect(
-      getByText(EN_TRANSLATIONS.createidentifier.threshold.subtitle)
-    ).toBeVisible();
-    expect(
-      getByText(EN_TRANSLATIONS.createidentifier.threshold.label)
-    ).toBeVisible();
-    expect(getByTestId("decrease-threshold-button")).toBeVisible();
-    expect(getByTestId("increase-threshold-button")).toBeVisible();
-    expect(
-      getByText(EN_TRANSLATIONS.createidentifier.threshold.continue)
-    ).toBeVisible();
+      getByText(
+        EN_TRANSLATIONS.createidentifier.connections.continue
+      ).getAttribute("disabled")
+    ).toBe("false");
+  });
+
+  test("Continue button disable status when select or unselect all connection", async () => {
+    const { getByTestId, getByText } = render(
+      <Provider store={storeMocked}>
+        <GroupMembers
+          state={stage2State}
+          setState={setState}
+          componentId={"create-identifier-modal"}
+          setBlur={setBlur}
+          resetModal={resetModal}
+        />
+      </Provider>
+    );
+
+    act(() => {
+      ionFireEvent.ionChange(getByTestId("connection-checkbox-0"), "false");
+    });
+
+    await waitFor(() => {
+      expect(
+        getByText(
+          EN_TRANSLATIONS.createidentifier.connections.continue
+        ).getAttribute("disabled")
+      ).toBe("true");
+    });
+
+    act(() => {
+      ionFireEvent.ionChange(getByTestId("connection-checkbox-0"), "false");
+    });
+
+    await waitFor(() => {
+      expect(
+        getByText(
+          EN_TRANSLATIONS.createidentifier.connections.continue
+        ).getAttribute("disabled")
+      ).toBe("false");
+    });
   });
 
   test("Continue button click", async () => {
     const { getByText } = render(
       <Provider store={storeMocked}>
-        <IdentifierStage3
-          state={stage3State}
+        <GroupMembers
+          state={stage2State}
           setState={setState}
           componentId={"create-identifier-modal"}
           setBlur={setBlur}
@@ -96,52 +133,23 @@ describe("Identifier Stage 3", () => {
 
     act(() => {
       fireEvent.click(
-        getByText(EN_TRANSLATIONS.createidentifier.threshold.continue)
+        getByText(EN_TRANSLATIONS.createidentifier.connections.continue)
       );
     });
 
     await waitFor(() => {
       expect(innerSetState).toBeCalledWith({
-        identifierCreationStage: 4,
+        identifierCreationStage: Stage.SetupThreshold,
+        selectedConnections: [connectionsFix[0]],
       });
     });
-  });
-
-  test("Increase/decrease threshold", async () => {
-    const { getByTestId } = render(
-      <Provider store={storeMocked}>
-        <IdentifierStage3
-          state={stage3State}
-          setState={setState}
-          componentId={"create-identifier-modal"}
-          setBlur={setBlur}
-          resetModal={resetModal}
-        />
-      </Provider>
-    );
-
-    act(() => {
-      fireEvent.click(getByTestId("increase-threshold-button"));
-    });
-
-    await waitFor(() => {
-      expect(innerSetState).toBeCalledWith({
-        threshold: 2,
-      });
-    });
-
-    act(() => {
-      fireEvent.click(getByTestId("decrease-threshold-button"));
-    });
-
-    expect(innerSetState).not.toHaveBeenNthCalledWith(2);
   });
 
   test("Back button click", async () => {
     const { getByText } = render(
       <Provider store={storeMocked}>
-        <IdentifierStage3
-          state={stage3State}
+        <GroupMembers
+          state={stage2State}
           setState={setState}
           componentId={"create-identifier-modal"}
           setBlur={setBlur}
@@ -156,8 +164,8 @@ describe("Identifier Stage 3", () => {
 
     await waitFor(() => {
       expect(innerSetState).toBeCalledWith({
-        identifierCreationStage: 2,
-        threshold: 1,
+        identifierCreationStage: Stage.SetupConnection,
+        selectedConnections: [],
       });
     });
   });
