+++ src/ui/components/CreateGroupIdentifier/index.ts
@@ -1,0 +1,1 @@
+export * from "./CreateGroupIdentifier";
