+++ src/ui/components/CreateIdentifier/components/IdentifierThemeSelector/IdentifierThemeSelector.scss
