+++ src/ui/components/CreateIdentifier/components/TypeItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./TypeItem";
