+++ src/ui/components/EditIdentifier/EditIdentifier.scss
@@ -31,10 +31,10 @@
       .indentifier-input {
         margin-bottom: 2.375rem;
         &.has-error {
+          margin-bottom: 0;
           ion-item.custom-input .input-line {
             border-color: var(--ion-color-danger);
           }
-          margin-bottom: 0;
         }
 
         .label-stacked {
