+++ src/ui/pages/Scan/Scan.test.tsx
@@ -171,6 +171,14 @@ describe("Scan Tab", () => {
         currentOperation: OperationType.IDLE,
         toastMsgs: [],
       },
+      identifiersCache: {
+        identifiers: [],
+        favourites: [],
+        multiSigGroup: {
+          groupId: "",
+          connections: [],
+        },
+      },
     };
 
     const storeMocked = {
